requirements.txt
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
idna==2.8
Kivy==1.10.1
Kivy-Garden==0.1.4
kivymd==0.98.5
Pillow==6.0.0
pygame==1.9.6
Pygments==2.3.1
requests==2.21.0
urllib3==1.24.2
